+++ ukvm/ukvm_hv_freebsd.c
@@ -130,3 +130,6 @@ struct ukvm_hv *ukvm_hv_init(size_t mem_size)
     hv->mem_size = mem_size;
     return hv;
 }
+
+void ukvm_hv_drop_privileges() {
+}
